+++ src/config/mod.rs
@@ -1,0 +1,2 @@
+pub mod schema;
+pub mod db_connection;
